+++ tests/nv/ex/test_scanner.py
@@ -508,6 +508,10 @@ class Test_scan_command(unittest.TestCase):
         self.assertRoute(['snoremap abc xyz', 'snor abc xyz'], cmd('snoremap', params={'lhs': 'abc', 'rhs': 'xyz'}))  # noqa: E501
         self.assertRoute(['snoremap', 'snor'], cmd('snoremap'))
         self.assertRoute(['sort', 'sor'], cmd('sort', addressable=True))
+        self.assertRoute(['sort i', 'sor i'], cmd('sort', params={'options': 'i'}, addressable=True))
+        self.assertRoute(['sort u', 'sor u'], cmd('sort', params={'options': 'u'}, addressable=True))
+        self.assertRoute(['sort ui', 'sor ui'], cmd('sort', params={'options': 'ui'}, addressable=True))
+        self.assertRoute(['sort iu', 'sor iu'], cmd('sort', params={'options': 'iu'}, addressable=True))
         self.assertRoute(['split file.txt', 'sp file.txt'], cmd('split', params={'file': 'file.txt'}))
         self.assertRoute(['split', 'sp'], cmd('split', params={'file': None}))
         self.assertRoute(['substitute', 's'], cmd('substitute', addressable=True))